Generation schema plugin development (NuoDB type mappings updated), dump\load fixes

+++ core/src/main/java/com/nuodb/migration/jdbc/type/jdbc2/JdbcTinyIntType.java
@@ -28,7 +28,6 @@
 package com.nuodb.migration.jdbc.type.jdbc2;
 
 import com.nuodb.migration.jdbc.type.JdbcType;
-import com.nuodb.migration.jdbc.type.JdbcTypeDesc;
 
 import java.sql.Types;
 
@@ -40,6 +39,6 @@ public class JdbcTinyIntType extends JdbcSmallIntType {
     public static final JdbcType INSTANCE = new JdbcTinyIntType();
 
     public JdbcTinyIntType() {
-        super(Types.SMALLINT);
+        super(Types.TINYINT);
     }
 }
